    \name{peakbound}
          \Rdversion{1.0}
          \alias{peakbound}
          \title{Return Exact Peak Bounds}
          \description{
            Uses basecounts generated from basecountimport() to determine exact boundaries of peaks from basecounts imported from significant windows. 
          }
          \usage{peakbound(bpprofile,output}
          \arguments{
            \item{bpprofile}{path to basecount file from basecountimport()}
            \item{output}{file that outputs start and stop positions with refined peaks, in addition to the location of the maximum of the peak and the basecount value at the maximum}
          }
          \seealso{
            \code{\link{save}}.
          }
          \examples{

   
          }
          \keyword{file}

