    \name{basecountimport}
          \Rdversion{1.0}
          \alias{basecountimport}
          \title{Import SBC (basecounts) Given Coordinate File}
          \description{
            Calculates the SBC (basecounts) from a specified basecount file for a given set of merged significant windows, given a threshold level and the results from getsigwindows().}
          \usage{basecountimport(inputfile,winlist,threshold=.01,method='pscl',printFullOut=0,
    outputfile,twobitfile,chromosome='all'}
          \arguments{
	    \item{inputfile}{path to basecount track containing SBPC information for the entire genome, generated by basealigncount()}
	    \item{winlist}{path to files outputted from getsigwindows()}
	    \item{outputfile}{path to where file of extracted basecounts should be placed}
            \item{chromosome}{which chromosomes in coordinate file to get basecounts for.  Default is all.}
	    \item{threshold}{ threshold for signfiicant windows corresponding to method used (mixture vs pscl)}
	    \item{printFullOut}{specifies the file format ouputted from getsigwindows, where 1 correspondinds to output of full dataset with posterior enrichment probability results, 0 corresponds to only genomic coordinates and posterior probabilities}
	    \item{method}{method used to generate files in winlist, either mixture or pscl}
	    \item{twobitfile}{path to current build of human genome in .2bit format}
	    \item{outputfile}{specify file name/path for temporary file that holds matrix of overlap information generated for peak refinement}
          }
          \seealso{
            \code{\link{save}}.
          }
          \examples{

   
          }
          \keyword{file}
